Upload the entire lightbox staff after the main site code
